Improve the way the unlink command removes links to packages
At the moment, when `zulu unlink` is run (either alone, or as part of
`zulu uninstall`), it loops through the files in the bin, share and
index objects in the index file, and removes each one. This works while
packages are static, but causes issues when the index is updated. If a
linked file is removed from the index, when that package is
unlinked/uninstalled, the file that was removed from the index will
remain in the bin/share/init directory as zulu no longer knows about it.

This new version of the unlink command searches each of the directories
for links to the package's source, so that all links will be removed,
regardless of whether or not they are still mentioned in the index.