Use zulu bundle --dump to write to packagefile during package install/uninstall
Previously, they would write to the packagefile manually, which could
lead to merge conflicts and duplication when the packagefile is synced
between computers.

By using the output of `zulu bundle --dump` we can ensure that the
packagefile contains any branch/tag information for the package, and is
consistently in alphabetical order to reduce conflicts.

Fix #78