Enforce builtins and important system commands
It's fairly common for people to create aliases to set preferred options
for import system commands like `git`, `grep`, `ls` etc. Since Zulu
relies heavily on the default functionality of such commands, and is
loaded fully into the environment, a simple alias can have a
catastrophic effect on Zulu's functionality. To try and get around this,
this commit attempts to prefix *all* ZSH builtins and UNIX default
commands.

Since the user interface of Zulu will not change, this will be released
in a patch release to fix the immediate bugs. HOWEVER, such a
fundamental change is highly likely to introduce bugs, so this will sit
in Zulu Next for a while before release to give people a chance to test
it properly.

Please update this PR with any bugs/comments/feedback rather than
raising new issues.